+++ src/RX8900RTC.h
@@ -75,24 +75,24 @@ typedef enum {
 
 
 typedef struct {
-  byte SEC     = 0x00;//0x00-0x59 (bcd)
-  byte MIN     = 0x00;//0x00-0x59 (bcd)
-  byte HOUR    = 0x00;//0x00-0x23 (bcd)
-  byte WEEK    = 0x01;//SUN=0x01,MON=0x02,TUE=0x04,WED=0x08,THU=0x10,FRI=0x20,SAT=0x40
-  byte DAY     = 0x01;//0x00-0x28-0x31)
-  byte MONTH   = 0x01;//JUN=0x01,FEB=0x02,MAR=0x03,APR=0x04,MAY=0x05,JUN=0x06,JLY=0x07,AUG=0x08,SEP=0x09,OCT=0x10,NOV=0x11,DEC=0x12
-  byte YEAR    = 0x00;//0x00-0x99
-  byte RAM;
-  byte MIN_Alarm;
-  byte HOUR_Alarm;
-  byte WEEK_DAY_Alarm;
-  byte Timer_Counter_0;
-  byte Timer_Counter_1;
-  byte Extension_Register;
-  byte Flag_Register;
-  byte Control_Register;
-  byte TEMP;
-  byte Backup_Function;
+  uint8_t SEC     = 0x00;//0x00-0x59 (bcd)
+  uint8_t MIN     = 0x00;//0x00-0x59 (bcd)
+  uint8_t HOUR    = 0x00;//0x00-0x23 (bcd)
+  uint8_t WEEK    = 0x01;//SUN=0x01,MON=0x02,TUE=0x04,WED=0x08,THU=0x10,FRI=0x20,SAT=0x40
+  uint8_t DAY     = 0x01;//0x00-0x28-0x31)
+  uint8_t MONTH   = 0x01;//JUN=0x01,FEB=0x02,MAR=0x03,APR=0x04,MAY=0x05,JUN=0x06,JLY=0x07,AUG=0x08,SEP=0x09,OCT=0x10,NOV=0x11,DEC=0x12
+  uint8_t YEAR    = 0x00;//0x00-0x99
+  uint8_t RAM;
+  uint8_t MIN_Alarm;
+  uint8_t HOUR_Alarm;
+  uint8_t WEEK_DAY_Alarm;
+  uint8_t Timer_Counter_0;
+  uint8_t Timer_Counter_1;
+  uint8_t Extension_Register;
+  uint8_t Flag_Register;
+  uint8_t Control_Register;
+  uint8_t TEMP;
+  uint8_t Backup_Function;
 } regRX8900_t;
 
 
@@ -103,18 +103,18 @@ class RX8900RTC {
     void init(void);
     void begin(void);
     static time_t get(void);    //must be static to work with setSyncProvider() in the Time library
-    byte set(time_t t);
+    uint8_t set(time_t t);
     static tmElements_t read(void);
-    byte write(tmElements_t tm);
-    void setFullAlarm(WEEK_DAY_ALARM_TYPES_t wdAlarmType, byte minute, byte hour, byte daydate);
-    void setAlarm(byte minute, byte hour);
-    void setAlarm(byte minute);
-    void setDayAlarm(byte minute, byte hour, byte daydate);   // set day of a month to daydate.
-    void setWeekAlarm(byte minute, byte hour, byte daydate);  // set week alarm mask to daydate such as SUN | SAT.
+    uint8_t write(tmElements_t tm);
+    void setFullAlarm(WEEK_DAY_ALARM_TYPES_t wdAlarmType, uint8_t minute, uint8_t hour, uint8_t daydate);
+    void setAlarm(uint8_t minute, uint8_t hour);
+    void setAlarm(uint8_t minute);
+    void setDayAlarm(uint8_t minute, uint8_t hour, uint8_t daydate);   // set day of a month to daydate.
+    void setWeekAlarm(uint8_t minute, uint8_t hour, uint8_t daydate);  // set week alarm mask to daydate such as SUN | SAT.
     void disableAlarm();
     void alarmInterrupt(INTERRUPT_CONTROL_t interrupt);
     bool alarm(void);  // Returns AF (Alarm Flag) status and reset AF if AF is "1".
-    void setFixedCycleTimer(int timerCounter, SOURCE_CLOCK_TYPES_t sourceCycle);
+    void setFixedCycleTimer(uint16_t timerCounter, SOURCE_CLOCK_TYPES_t sourceCycle);
     void disableFixedCycleTimer(void);
     void fixedCycleTimerInterrupt(INTERRUPT_CONTROL_t interrupt);
     bool fixedCycleTimer(void); // Returns TF (Timer Flag) status and reset TF if TF is "1".
@@ -128,14 +128,14 @@ class RX8900RTC {
   private:
 
     void COUNT_START(void);
-    void ByteWrite(byte reg, byte data);
-    byte ByteRead(byte reg);
+    void ByteWrite(uint8_t reg, uint8_t data);
+    uint8_t ByteRead(uint8_t reg);
     static regRX8900_t RegisterRead(void);
     void RESET(void);
     void SET_AIE(void);              //ALARM INTERRUPT ENABLE
     void RESET_AIE(void);            //ALARM INTERRUPT DISABLE
     void RESET_AF(void);             //RESET ALARM FLAG
-    byte IS_AF(void);                //IS ALARM TIME?
+    uint8_t IS_AF(void);                //IS ALARM TIME?
     void SET_AE(ALARM_TYPES_t reg);  //ALARM ENABLE
     void RESET_AE(ALARM_TYPES_t reg);//ALARM DISABLE
     void SET_WADA_W(void);           //SET WADA BIT 0 (WEEK ALARM)
@@ -144,7 +144,7 @@ class RX8900RTC {
     void SET_TIE(void);              //TIMER INTERRUPT ENABLE
     void RESET_TIE(void);            //TIMER INTERRUPT DISABLE
     void RESET_TF(void);             //RESET TIMER FLAG
-    byte IS_TF(void);                //IS TIMER COUNT UP?
+    uint8_t IS_TF(void);                //IS TIMER COUNT UP?
     void SET_TE(void);               //TIMER ENABLE
     void RESET_TE(void);             //TIMER DISABLE
     void SET_TSEL_1S(void);          //SET TIMER INTERVAL 1sec
@@ -153,12 +153,12 @@ class RX8900RTC {
     void SET_UIE(void);              //UPDATE INTERRUPT ENABLE
     void RESET_UIE(void);            //UPDATE INTERRUPT DISABLE
     void RESET_UF(void);             //RESET UPDATE FLAG
-    byte IS_UF(void);                //IS UPDATE INTERRUPT?
+    uint8_t IS_UF(void);                //IS UPDATE INTERRUPT?
     void SET_USEL(USEL_t ust);       //SET UPDATE INTERRUPT SELECT
 
-    byte subZeller( int y, int m, int d );
-    byte dec2bcd(byte n);
-    static byte bcd2dec(byte n);
+    uint8_t subZeller(uint16_t y, uint16_t m, uint16_t d);
+    uint8_t dec2bcd(uint8_t n);
+    static uint8_t bcd2dec(uint8_t n);
 };
 
 //extern RX8900RTC RTC;
